Tests/LibWeb/Text/expected/navigation/history-popstate-event.txt

popstate event from iframe new_path=iframe-popstate-event.html
